Add account role list command
Add new command for account: account role list
