+++ pkg/consts/endpoint/endpoint.go
@@ -21,3 +21,4 @@ var AvailableRatePlanEndpoint = "/zones/%s/available_rate_plans"
 var AccountListEndPoint = "/accounts?per_page=50&direction=desc"
 var AccountDetailsEndpoint = "/accounts/%s"
 var AccountUsersEndpoint = "/accounts/%s/members"
+var AccountRolesEndpoint = "/accounts/%s/roles"
